move download-assets to utils
