+++ src/utils/download-assets.js
@@ -95,4 +95,4 @@ export function cleanup(folder) {
       }
     })
   }
-}
+}
